Moved footer and footer styling to folder. Adjusted coloring and bordering for theme consistency
